started writing a bat file which automates the task of generating release package for the engine. This release package includes: The Dll of the engine in Debug and Release mode The Includes of the engine The SandBox.cpp file as starter scene res/engine folder with all the engine resources in it res/game folder for all the starter content in it premake5 exe and license premake5.lua file for project generation GenerateProjectFile.bat file for create vs2017 project files with the lua file.
the above all need to be in the release package folder. I have only implmented a few on the list
the rest still need to be done.

also change the lua file for the main engine solution to not copy over
the include files to the ToyRenderer binaray folder since this new batch
file is responisible for that